fix: GX2F: update track state parameters and converge (#1702)
The previous implementation could not update the track state parameters.
The updated Jacobian was overridden by `stepper.transportCovarianceToBound`. Therefore, every time the identity matrix was used in place of the Jacobian.
A second problem was the handling of surfaces in the update steps. Now, during the first iteration, proxies are created for each surface. During further iterations those surfaces are visited again.