build: Git commit hash is retrieved at install time (#4673)
Currently, at configure time, we will generate a header file containing the git commit hash. This means that when the commit changes, cmake reconfigures and the core lib is recompiled. With this change, the absence of a commit hash is explicitly handled with `std::optional`. At install time, we look up the commit hash and regenerate the header file into the install directory. If the source folder is not a repository (like when building from a tarball), the configure-time header is copied with the `nullopt` commit hash to indicate not commit is available.